cmdlib: Move to cosalib
